+++ library/Dlayer/DesignerTool/FormBuilder/Text/SubTool/Styling/Tool.php
@@ -7,7 +7,8 @@
  * @copyright G3D Development Limited
  * @license https://github.com/Dlayer/dlayer/blob/master/LICENSE
  */
-class Dlayer_DesignerTool_FormBuilder_Text_Tool extends Dlayer_DesignerTool_FormBuilder_Shared_Tool_Styling
+class Dlayer_DesignerTool_FormBuilder_Text_SubTool_Styling_Tool extends
+    Dlayer_DesignerTool_FormBuilder_Shared_Tool_Styling
 {
     // No custom logic
 }