+++ library/Dlayer/Ribbon/Handler/Form.php
@@ -126,6 +126,11 @@ class Dlayer_Ribbon_Handler_Form
                 $data = $ribbon_text->viewData($this->toolParams($tool));
                 break;
 
+            case 'styling':
+                $ribbon_text = new Dlayer_DesignerTool_FormBuilder_Text_SubTool_Styling_Ribbon();
+                $data = $ribbon_text->viewData($this->toolParams($tool));
+                break;
+
             default:
                 $data = false;
                 break;
@@ -148,6 +153,11 @@ class Dlayer_Ribbon_Handler_Form
         switch ($tab) {
             case 'textarea':
                 $ribbon_textarea = new Dlayer_DesignerTool_FormBuilder_Textarea_Ribbon();
+                $data = $ribbon_textarea->viewData($this->toolParams($tool));
+                break;
+
+            case 'styling':
+                $ribbon_textarea = new Dlayer_DesignerTool_FormBuilder_Textarea_SubTool_Styling_Ribbon();
                 $data = $ribbon_textarea->viewData($this->toolParams($tool));
                 break;
 
@@ -176,6 +186,11 @@ class Dlayer_Ribbon_Handler_Form
                 $data = $ribbon_email->viewData($this->toolParams($tool));
                 break;
 
+            case 'styling':
+                $ribbon_email = new Dlayer_DesignerTool_FormBuilder_Email_SubTool_Styling_Ribbon();
+                $data = $ribbon_email->viewData($this->toolParams($tool));
+                break;
+
             default:
                 $data = false;
                 break;
@@ -198,6 +213,11 @@ class Dlayer_Ribbon_Handler_Form
         switch ($tab) {
             case 'password':
                 $ribbon_password = new Dlayer_DesignerTool_FormBuilder_Password_Ribbon();
+                $data = $ribbon_password->viewData($this->toolParams($tool));
+                break;
+
+            case 'styling':
+                $ribbon_password = new Dlayer_DesignerTool_FormBuilder_Password_SubTool_Styling_Ribbon();
                 $data = $ribbon_password->viewData($this->toolParams($tool));
                 break;