+++ private/database/tables/data/dlayer_module_tool_tab.sql
@@ -40,10 +40,14 @@ insert  into `dlayer_module_tool_tab`(`id`,`module_id`,`tool_id`,`name`,`script`
     (NULL,4,10,'Settings','settings','Settings','cog',NULL,1,0,0,2,1),
     (NULL,4,10,'Responsive','responsive','Responsive','equalizer',"transform: rotate(90deg);",1,0,0,3,1),
     (NULL,3,14,'Text','text',NULL,'pencil',NULL,1,0,1,1,1),
-    (NULL,3,14,'Help','help',NULL,'info-sign',NULL,0,0,0,2,1),
+    (NULL,3,14,'Help','help',NULL,'info-sign',NULL,0,0,0,3,1),
     (NULL,3,15,'Textarea','textarea',NULL,'pencil',NULL,1,0,1,1,1),
-    (NULL,3,15,'Help','help',NULL,'info-sign',NULL,0,0,0,2,1),
+    (NULL,3,15,'Help','help',NULL,'info-sign',NULL,0,0,0,3,1),
     (NULL,3,16,'Email','email',NULL,'pencil',NULL,1,0,1,1,1),
-    (NULL,3,16,'Help','help',NULL,'info-sign',NULL,0,0,0,2,1),
+    (NULL,3,16,'Help','help',NULL,'info-sign',NULL,0,0,0,3,1),
     (NULL,3,17,'Password','password',NULL,'pencil',NULL,1,0,1,1,1),
-    (NULL,3,17,'Help','help',NULL,'info-sign',NULL,0,0,0,2,1);
+    (NULL,3,17,'Help','help',NULL,'info-sign',NULL,0,0,0,3,1),
+    (NULL,3,14,'Styling','styling','Styling','tint',NULL,1,1,0,2,1),
+    (NULL,3,15,'Styling','styling','Styling','tint',NULL,1,1,0,2,1),
+    (NULL,3,16,'Styling','styling','Styling','tint',NULL,1,1,0,2,1),
+    (NULL,3,17,'Styling','styling','Styling','tint',NULL,1,1,0,2,1);